Refactor into separate binder/checker.
