+++ adl/language/compiler/parser.ts
@@ -54,7 +54,7 @@ export function parse(code: string) {
           return parseAliasStatement();
       }
 
-      throw new Error('Unable to parse statement');
+      throw error(`Expected statement, but found ${Kind[tok]}`);
     }
 
   }
@@ -499,7 +499,7 @@ export function parse(code: string) {
   }
 
   function error(msg: string) {
-    throw new Error(`[${scaner.position.line}, ${scaner.position.character}] ${msg}`);
+    throw new Error(`[${scaner.position.line + 1}, ${scaner.position.character + 1}] ${msg}`);
   }
 
   function parseExpected(expectedToken: Kind) {
