+++ adl/language/samples/petstore/decorators.js
@@ -1,9 +1,9 @@
 // pretend I just typed this file from scratch... and saved it
 import { description } from "../../lib/decorators.js";
 
-export function fancyDescription(program, target, ...args) {
-  args[0] = `<blink>${args[0]}</blink>`;
-  description(program, target, ...args);
+export function fancyDescription(program, target, text) {
+  text = `<blink>${text}</blink>`;
+  description(program, target, text);
 }
 
 export function evenFancierDescription(program, target, ...args) {
